refactor: 결과 출력 흐름 report 관련 객체로 이동
