+++ src/main/java/controller/GameMode.java
@@ -10,6 +10,7 @@ import model.judgement.Judgement;
 import model.judgement.PlayerResult;
 import model.paticipant.Dealer;
 import model.paticipant.Players;
+import view.OutputView;
 
 public class GameMode {
 
@@ -34,7 +35,15 @@ public class GameMode {
     }
 
     public void reportResult(Dealer dealer, Players players) {
+        printFinalCards(dealer, players);
+
         PlayerResult playerResult = Judgement.judgeByPlayer(dealer, players);
         resultReporter.report(playerResult);
+    }
+
+    private void printFinalCards(Dealer dealer, Players players) {
+        OutputView.printBlank();
+        OutputView.printCardByPlayerWithScore(dealer);
+        players.forEach(OutputView::printCardByPlayerWithScore);
     }
 }
